requirements.txt
# TranscribeAI - Complete Dependencies List
# Install with: pip install -r requirements.txt

# ===== Web Framework =====
Flask>=2.3.0
flask-cors>=4.0.0
Werkzeug>=3.0.0

# ===== Audio Processing =====
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
resampy>=0.4.0

# ===== Numerical Computing =====
numpy>=1.26.0
scipy>=1.11.0
numba>=0.58.0

# ===== Machine Learning (Optional but recommended) =====
torch>=2.0.0
scikit-learn>=1.3.0

# ===== MIDI and Music Notation =====
pretty_midi>=0.2.10
music21>=9.1.0

# ===== Data Processing and Utilities =====
PyYAML>=6.0
joblib>=1.3.0
matplotlib>=3.7.0
pandas>=2.1.0
tqdm>=4.66.0

# ===== Jupyter Notebook Support (Optional) =====
jupyter>=1.0.0
notebook>=7.0.0
ipython>=8.12.0

# ===== Additional Audio Dependencies =====
ffmpeg-python>=0.2.0
pydub>=0.25.0

# ===== System Dependencies =====
# Note: You may also need to install ffmpeg separately:
# - Windows: choco install ffmpeg  OR  download from https://ffmpeg.org/
# - macOS: brew install ffmpeg
# - Linux: sudo apt-get install ffmpeg
